clean up and organize stories
