+++ .storybook/preview.js
@@ -7,6 +7,11 @@ const preview = {
         color: /(background|color)$/i,
         date: /Date$/i,
       },
+      options: {
+        storySort: {
+          order: ["Pages", "Components"],
+        },
+      },
     },
   },
 };
