--- components/HamburgerMenu.tsx
@@ -1,53 +1,0 @@
-import Link from "next/link";
-import React, { useState } from "react";
-
-const HamburgerMenu = () => {
-  const [isOpen, setIsOpen] = useState(false);
-
-  const toggleMenu = () => {
-    setIsOpen(!isOpen);
-  };
-
-  return (
-    <div>
-      <button
-        onClick={toggleMenu}
-        style={{
-          position: "absolute",
-          top: "10px",
-          left: "10px",
-          zIndex: 1000,
-        }}
-      >
-        &#9776;
-      </button>
-      {isOpen && (
-        <div
-          style={{
-            position: "absolute",
-            top: "40px",
-            left: "10px",
-            backgroundColor: "white",
-            border: "1px solid #ccc",
-            borderRadius: "5px",
-            zIndex: 1000,
-          }}
-        >
-          <ul style={{ listStyleType: "none", padding: "10px" }}>
-            <li>
-              <Link href="/">Home</Link>
-            </li>
-            <li>
-              <Link href="/profile">Profile</Link>
-            </li>
-            <li>
-              <Link href="/settings">Settings</Link>
-            </li>
-          </ul>
-        </div>
-      )}
-    </div>
-  );
-};
-
-export default HamburgerMenu;
